refactoring apollo-server to run on next pages via api/graphql, installing as-integrations/next and graphql-tag, removing server folder from top level directory as next already provides a server and code now leverages existing server to have an endpoint for graphql apis to be called on
