+++ package.json
@@ -8,8 +8,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "compile": "tsc",
-    "server": "npm run compile && ts-node --esm ./server/server.ts"
+    "compile": "tsc"
   },
   "repository": {
     "type": "git",
@@ -24,7 +23,9 @@
   "homepage": "https://github.com/richardjhong/task-mate#readme",
   "dependencies": {
     "@apollo/server": "^4.6.0",
+    "@as-integrations/next": "^1.3.0",
     "graphql": "^16.6.0",
+    "graphql-tag": "^2.12.6",
     "next": "^13.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
@@ -32,6 +33,7 @@
   "devDependencies": {
     "@types/node": "^18.15.11",
     "@types/react": "^18.0.33",
+    "concurrently": "^8.0.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
   },
