+++ src/pages/api/graphql.ts
@@ -1,0 +1,10 @@
+import { ApolloServer } from '@apollo/server';
+import { startServerAndCreateNextHandler } from '@as-integrations/next';
+import { typeDefs, resolvers } from './schemas/index.ts';
+
+const server = new ApolloServer({
+  resolvers,
+  typeDefs,
+});
+
+export default startServerAndCreateNextHandler(server);
