+++ src/pages/api/schemas/typeDefs.ts
@@ -1,0 +1,40 @@
+
+// // A schema is a collection of type definitions (hence "typeDefs")
+// // that together define the "shape" of queries that are executed against
+// // your data.
+// const typeDefs = `#graphql
+//   # Comments in GraphQL strings (such as this one) start with the hash (#) symbol.
+
+//   # This "Book" type defines the queryable fields for every book in our data source.
+//   type Book {
+//     title: String
+//     author: String
+//   }
+
+//   # The "Query" type is special: it lists all of the available queries that
+//   # clients can execute, along with the return type for each. In this
+//   # case, the "books" query returns an array of zero or more Books (defined above).
+//   type Query {
+//     books: [Book]
+//   }
+// `;
+
+// export { typeDefs };
+
+import { gql } from 'graphql-tag'
+
+const typeDefs = gql`
+    type Query {
+      books: [Book]
+    }
+
+  type Book {
+    title: String
+    author: String
+  }
+  
+  type Example {
+    name: String!
+  }
+`
+export { typeDefs };
